+++ ktor/common/src/jsMain/kotlin/dev/inmo/micro_utils/ktor/common/ActualMPPInput.kt
@@ -1,0 +1,7 @@
+package dev.inmo.micro_utils.ktor.common
+
+import dev.inmo.micro_utils.common.*
+import io.ktor.utils.io.core.ByteReadPacket
+import io.ktor.utils.io.core.Input
+
+actual fun MPPFile.input(): Input = ByteReadPacket(bytesSync())